src/requirements.txt
google.auth
google-cloud-resource-manager
requests